Move constant templates to ui
