+++ src/ui/registry/units/editor/register.tsx
@@ -1,7 +1,7 @@
-import {getEditorTemplates} from '../../../../shared/constants/editor-templates';
 import {registry} from '../../../registry';
 import {exampleFunction} from '../../../registry/functions/example-function';
 import {EXAMPLE_FUNCTION} from '../../../registry/units/common/constants/functions';
+import {getEditorTemplates} from '../../../units/editor/constants/templates';
 
 export const registerEditorPlugins = () => {
     registry.editor.functions.register({
